# api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Aplikacje Mobilne API",
    "description": "Aplikacje Mobilne API made with Express.js",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "tags": ["root"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/transactions/list": {
      "get": {
        "description": "",
        "parameters": [],
        "tags": ["transaction"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "description": "",
        "tags": ["transaction"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["transaction"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["transaction"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                },
                "categoryId": {
                  "example": "any"
                },
                "currencyId": {
                  "example": "any"
                },
                "typeId": {
                  "example": "any"
                },
                "priorityId": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/transactions/": {
      "get": {
        "description": "",
        "tags": ["transaction"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["transaction"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                },
                "categoryId": {
                  "example": "any"
                },
                "currencyId": {
                  "example": "any"
                },
                "typeId": {
                  "example": "any"
                },
                "priorityId": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "description": "",
        "tags": ["category"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["category"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["category"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/categories/name/{name}": {
      "get": {
        "description": "",
        "tags": ["category"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/categories/": {
      "get": {
        "description": "",
        "tags": ["category"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["category"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/currencies/{id}": {
      "get": {
        "description": "",
        "tags": ["currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/currencies/": {
      "get": {
        "description": "",
        "parameters": [],
        "tags": ["currency"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["currency"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "symbol": {
                  "example": "any"
                },
                "acronym": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/types/{id}": {
      "get": {
        "description": "",
        "tags": ["type"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/types/": {
      "get": {
        "description": "",
        "parameters": [],
        "tags": ["type"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/priorities/{id}": {
      "get": {
        "description": "",
        "tags": ["priority"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/priorities/": {
      "get": {
        "description": "",
        "parameters": [],
        "tags": ["priority"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/priorities/name/{name}": {
      "get": {
        "description": "",
        "tags": ["priority"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    
"/api/attachments/": {
  "get": {
    "description": "",
    "parameters": [],
    "tags": ["attachment"],
    "responses": {
      "200": {
        "description": "OK"
      },
      "400": {
        "description": "Bad Request"
      }
    }
  },
  "post": {
    "description": "",
    "parameters": [
      {
        "url": "body",
        "transactionId": "body",
        "schema": {
          "type": "object",
          "properties": {
            "url": {
              "example": "any"
            },
            "transactionId": {
              "example": "any"
            }

          }
        }
      }
    ],
    "tags": ["attachment"],
    "responses": {
      "200": {
        "description": "OK"
      },
      "400": {
        "description": "Bad Request"
      }
    }
  },
  "delete": {
    "description": "",
    "parameters": [],
    "tags": ["attachment"],
    "responses": {
      "200": {
        "description": "OK"
      },
      "400": {
        "description": "Bad Request"
      }
    }
  }
},
"/api/attachments/{id}": {
  "get": {
    "description": "",
    "tags": ["attachment"],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "type": "string"
      }
    ],
    "responses": {
      "200": {
        "description": "OK"
      },
      "400": {
        "description": "Bad Request"
      }
    }
  }
}
  }
}